more blog formatting and construction
