+++ components/blog/BlogHomeSection.tsx
@@ -1,6 +1,7 @@
 import Image from "next/image";
 import Link from "next/link";
 import { PostMeta } from "../../lib/types/PostMeta";
+import { BlogPost } from "/types";
 
 interface Props {
 	posts: BlogPost[];
