+++ package.json
@@ -30,6 +30,7 @@
 		"next-sanity": "^0.5.1",
 		"next-sanity-image": "^3.2.1",
 		"react": "17.0.2",
-		"react-dom": "17.0.2"
+		"react-dom": "17.0.2",
+		"@portabletext/react": "^1.0.4"
 	}
 }
